+++ packages/talker_flutter/lib/src/extensions/log_level.dart
@@ -6,18 +6,19 @@ extension ToColor on LogLevel? {
   Color get color {
     switch (this) {
       case LogLevel.critical:
+        return Colors.red[800]!;
       case LogLevel.error:
-        return Colors.red;
+        return Colors.red[400]!;
       case LogLevel.fine:
-        return Colors.teal;
+        return Colors.teal[400]!;
       case LogLevel.warning:
-        return Colors.orange;
+        return Colors.orange[800]!;
       case LogLevel.verbose:
-        return Colors.grey;
+        return Colors.grey[400]!;
       case LogLevel.info:
-        return Colors.blue;
+        return Colors.blue[400]!;
       case LogLevel.good:
-        return Colors.green;
+        return Colors.green[400]!;
       case LogLevel.debug:
       default:
         return Colors.grey;