+++ packages/talker_flutter/lib/src/src.dart
@@ -1,4 +1,4 @@
 export 'talker_screen.dart';
 export 'extensions/extensions.dart';
 export 'models/models.dart';
-export 'options/options.dart';
+export 'theme/theme.dart';